requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
algosdk==2.7.0
python-dotenv==1.0.0
google-generativeai==0.8.5
langgraph==0.0.29
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-ansible-filters==1.3.2
joblib==1.5.2
jsondiff==2.2.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
langchain-core==1.0.5
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mslex==1.3.0
multiformats==0.3.1
multiformats-config==0.3.1
nest-asyncio==1.6.0
networkx==3.5
numpy==1.24.3
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.0.3
parso==0.8.5
pathspec==0.12.1
pipx==1.8.0
platformdirs==4.5.0
plumbum==1.9.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
py-algorand-sdk==2.11.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclip==0.7.0
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
pyteal==0.27.0
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.1.0
questionary==1.10.0
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scapy==2.6.1
scikit-learn==1.3.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textual==3.7.1
threadpoolctl==3.6.0
tokenizers==0.15.0
torch==2.0.1
tornado==6.5.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.2
typing-inspection==0.4.2
typing-validation==1.2.12
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
userpath==1.9.2
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
